Added Loading ROM code
Added the loading of a rom, not working so far

+++ Wrapper/RetroPortableUI.Wrapper.LibRetro/RetroStructures/RetroGameInfo.cs
@@ -8,7 +8,7 @@ namespace RetroPortableUI.Wrapper.LibRetro.RetroStructures
     {
         public char* path;
         public void* data;
-        public byte size;
+        public uint size;
         public char* meta;
     }
 }
